Allow support for older Chef Infra server singin requests
