+++ chef-guard.go
@@ -38,7 +38,7 @@ import (
 )
 
 // VERSION holds the current version
-const VERSION = "0.7.3"
+const VERSION = "0.8.0"
 
 var chefKey string
