+++ cookbook.go
@@ -22,7 +22,8 @@ import (
 	"compress/gzip"
 	"crypto/hmac"
 	"crypto/md5"
-	"encoding/hex"
+	"crypto/sha1"
+	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -34,14 +35,12 @@ import (
 	"os"
 	"path"
 	"regexp"
-	"sort"
 	"strings"
 	"time"
 
 	// import hmac
 
 	// import hashlib
-	"crypto/sha256"
 
 	"github.com/gorilla/mux"
 )
@@ -324,150 +323,33 @@ func downloadCookbookFile(c *http.Client, orgID, checksum string) ([]byte, error
 	return ioutil.ReadAll(resp.Body)
 }
 
-// HMAC SHA256
-func hmacSHA256(key, data []byte) []byte {
-	h := hmac.New(sha256.New, key)
-	h.Write(data)
-	return h.Sum(nil)
-}
+func generateV12SignedURL(orgID, checksum string) (*url.URL, error) {
+	expires := time.Now().Unix() + 10
+	stringToSign := fmt.Sprintf("GET\n\n\n%d\n/bookshelf/organization-%s/checksum-%s", expires, orgID, checksum)
 
-// Hash SHA256
-func hashSHA256(data string) string {
-	h := sha256.New()
-	h.Write([]byte(data))
-	return hex.EncodeToString(h.Sum(nil))
-}
+	h := hmac.New(sha1.New, []byte(cfg.Chef.BookshelfSecret))
+	h.Write([]byte(stringToSign))
+	signature := url.QueryEscape(base64.StdEncoding.EncodeToString(h.Sum(nil)))
 
-// Create the signing key
-func getSignatureKey(secret, date, region, service string) []byte {
-	kDate := hmacSHA256([]byte("AWS4"+secret), []byte(date))
-	kRegion := hmacSHA256(kDate, []byte(region))
-	kService := hmacSHA256(kRegion, []byte(service))
-	kSigning := hmacSHA256(kService, []byte("aws4_request"))
-	return kSigning
-}
-
-// Generate the canonical query string
-func getCanonicalQueryString(params url.Values) string {
-	keys := make([]string, 0, len(params))
-	for k := range params {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-
-	queryString := make([]string, 0, len(params))
-	for _, k := range keys {
-		queryString = append(queryString, fmt.Sprintf("%s=%s", url.QueryEscape(k), url.QueryEscape(params.Get(k))))
-	}
-
-	return strings.Join(queryString, "&")
-}
-
-// Get canonical headers and signed headers
-func getCanonicalHeaders(headers map[string]string) (string, string) {
-	var headerKeys []string
-	for k := range headers {
-		headerKeys = append(headerKeys, strings.ToLower(k))
-	}
-	sort.Strings(headerKeys)
-	var canonicalHeaders, signedHeaders string
-	for _, k := range headerKeys {
-		canonicalHeaders += fmt.Sprintf("%s:%s\n", k, strings.TrimSpace(headers[k]))
-		signedHeaders += fmt.Sprintf("%s;", k)
-	}
-	signedHeaders = strings.TrimSuffix(signedHeaders, ";")
-	return canonicalHeaders, signedHeaders
-}
-
-// Generate the signed URL
-func getSignedURL(accessKey, secretKey, region, service, method, uri string, queryParams map[string]string, headers map[string]string, payload string) string {
-	algorithm := "AWS4-HMAC-SHA256"
-	t := time.Now()
-	date := t.Format("20060102")
-	amzDate := t.Format("20060102T150405Z")
-	credentialScope := fmt.Sprintf("%s/%s/%s/aws4_request", date, region, service)
-
-	query := url.Values{}
-	for k, v := range queryParams {
-		query.Set(k, v)
-	}
-	query.Set("X-Amz-Algorithm", algorithm)
-	query.Set("X-Amz-Credential", fmt.Sprintf("%s/%s", accessKey, credentialScope))
-	query.Set("X-Amz-Date", amzDate)
-	query.Set("X-Amz-Expires", "86400")
-	query.Set("X-Amz-SignedHeaders", "host")
-
-	canonicalURI := uri
-	canonicalQueryString := getCanonicalQueryString(query)
-	canonicalHeaders, signedHeaders := getCanonicalHeaders(headers)
-
-	canonicalRequest := fmt.Sprintf("%s\n%s\n%s\n%s\n%s\n%s",
-		method,
-		canonicalURI,
-		canonicalQueryString,
-		canonicalHeaders,
-		signedHeaders,
-		payload,
+	urlStr := fmt.Sprintf(
+		"%s/bookshelf/organization-%s/checksum-%s?AWSAccessKeyId=%s&Expires=%d&Signature=%s",
+		getChefBaseURL(),
+		orgID,
+		checksum,
+		cfg.Chef.BookshelfKey,
+		expires,
+		signature,
 	)
 
-	// GET
-	// /bookshelf/organization-0751dc28b5a6978ca80465d54cc7f6ff/checksum-6d5e878f0b1e6b2c9e368a4e6234462b
-	// X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=00efc6e783b514a4516a%2F20240717%2Fchef%2Fs3%2Faws4_request&X-Amz-Date=20240717T151227Z&X-Amz-Expires=86400
-	// host:infra.chef.saas.acc.schubergphilis.com
-
-	// host
-	// UNSIGNED-PAYLOAD
-
-	hashedCanonicalRequest := hashSHA256(canonicalRequest)
-	stringToSign := fmt.Sprintf("%s\n%s\n%s\n%s",
-		algorithm,
-		amzDate,
-		credentialScope,
-		hashedCanonicalRequest,
-	)
-
-	// AWS4-HMAC-SHA256
-	// 20240717T171604Z
-	// 20240717/chef/s3/aws4_request
-	// bca48d5c547f6d8b1412a648e9b47a445bdc8173aa6cfe5e61830e8142c1e283
-
-	signingKey := getSignatureKey(secretKey, date, region, service)
-	signature := hex.EncodeToString(hmacSHA256(signingKey, []byte(stringToSign)))
-
-	query.Set("X-Amz-Signature", signature)
-
-	signedURL := fmt.Sprintf("https://%s%s?%s", headers["host"], uri, query.Encode())
-	return signedURL
+	return url.Parse(urlStr)
 }
 
 func generateSignedURL(orgID, checksum string) (*url.URL, error) {
-
-	accessKey := cfg.Chef.BookshelfKey
-	secretKey := cfg.Chef.BookshelfSecret
-	region := cfg.Chef.BookshelfRegion
-	service := "s3"
-	method := "GET"
-	uri := fmt.Sprintf("/%s/organization-%s/checksum-%s", cfg.Chef.BookshelfBucket, orgID, checksum)
-	bookshelfDomain := cfg.Chef.BookshelfDomain
-
-	queryParams := map[string]string{}
-
-	headers := map[string]string{
-		"host": bookshelfDomain,
+	if cfg.Chef.Version >= 15 {
+		return generateAWSV4SignedURL(orgID, checksum)
+	} else {
+		return generateV12SignedURL(orgID, checksum)
 	}
-
-	signedURL := getSignedURL(accessKey, secretKey, region, service, method, uri, queryParams, headers, "UNSIGNED-PAYLOAD")
-
-	// https://infra.chef.saas.acc.schubergphilis.com:443
-	// /bookshelf/organization-0751dc28b5a6978ca80465d54cc7f6ff/checksum-5a64b525b6b148539060365ee4980839
-	// ?X-Amz-Algorithm=AWS4-HMAC-SHA256
-	// &X-Amz-Credential=00efc6e783b514a4516a%2F20240717%2Fchef%2Fs3%2Faws4_request
-	// &X-Amz-Date=20240717T112134Z
-	// &X-Amz-Expires=10800
-	// &X-Amz-SignedHeaders=host
-	// &X-Amz-Signature=10ee4d844d505fa7d7d96133634f93982ba9f3d9e939ffd632201a339ad6244a
-
-	return url.Parse(signedURL)
 }
 
 func writeFileToDisk(filePath string, content io.Reader) error {
